TG-03: id valid

+++ src/controllers/employees.js
@@ -1,6 +1,16 @@
+import mongoose from 'mongoose';
 import employees from '../models/Employees';
 
 const updatedEmployee = async (req, res) => {
+    const employeeId = req.params.id;
+
+    if (req.params.id && !mongoose.Types.ObjectId.isValid(employeeId)) {
+        return res.status(404).json({
+            message: `Employee id ${employeeId} not valid`,
+            data: undefined,
+            error: true,
+        });
+    }
   try {
     const { id } = req.params;
     const result = await employees.findByIdAndUpdate(
@@ -30,28 +40,37 @@ const updatedEmployee = async (req, res) => {
 };
 
 const deletedEmployee = async (req, res) => {
-  try {
+    const employeeId = req.params.id;
+
+    if (req.params.id && !mongoose.Types.ObjectId.isValid(employeeId)) {
+        return res.status(404).json({
+            message: `Employee id ${employeeId} not valid`,
+            data: undefined,
+            error: true,
+        });
+    }
+    try {
     const { id } = req.params;
     const result = await employees.findByIdAndDelete(id);
 
-    if (result) {
-      return res.status(200).json({
-        message: 'Employee deleted succesfully',
-        data: result,
-        error: false,
-      });
-    }
-    return res.status(404).json({
-      message: `Employee with id ${id} not found`,
-      data: undefined,
-      error: true,
-    });
+        if (result) {
+        return res.status(200).json({
+            message: 'Employee deleted succesfully',
+            data: result,
+            error: false,
+        });
+        }
+        return res.status(404).json({
+        message: `Employee with id ${id} not found`,
+        data: undefined,
+        error: true,
+        });
   } catch (error) {
-    return res.status(500).json({
-      message: `Server error: ${error}`,
-      data: undefined,
-      error: true,
-    });
+        return res.status(500).json({
+        message: `Server error: ${error}`,
+        data: undefined,
+        error: true,
+        });
   }
 };
